requirements.txt
flask>=2.0.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.30.0
PyPDF2>=3.0.0
tqdm>=4.65.0
